Move ParsePortRange into portrange.go.
