+++ network/portrange.go
@@ -6,6 +6,7 @@ package network
 import (
 	"fmt"
 	"sort"
+	"strconv"
 	"strings"
 
 	"github.com/juju/errors"
@@ -128,4 +129,74 @@ func CollapsePorts(ports []Port) (result []PortRange) {
 
 	}
 	return
+}
+
+// ParsePortRange builds a PortRange from the provided string. If the
+// string does not include a protocol then "tcp" is used. Validate()
+// gets called on the result before returning. If validation fails the
+// invalid PortRange is still returned.
+// Example strings: "80/tcp", "443", "12345-12349/udp".
+func ParsePortRange(portRangeStr string) (*PortRange, error) {
+	// Extract the protocol.
+	protocol := "tcp"
+	parts := strings.SplitN(portRangeStr, "/", 2)
+	if len(parts) == 2 {
+		portRangeStr = parts[0]
+		protocol = parts[1]
+	}
+
+	// Parse the ports.
+	portRange, err := parsePortRange(portRangeStr)
+	if err != nil {
+		return nil, errors.Trace(err)
+	}
+	portRange.Protocol = protocol
+
+	return portRange, portRange.Validate()
+}
+
+func parsePortRange(portRangeStr string) (*PortRange, error) {
+	var start, end int
+	parts := strings.Split(portRangeStr, "-")
+	if len(parts) > 2 {
+		return nil, errors.Errorf("invalid port range %q", portRangeStr)
+	}
+
+	if len(parts) == 1 {
+		port, err := strconv.Atoi(parts[0])
+		if err != nil {
+			return nil, errors.Annotatef(err, "invalid port %q", portRangeStr)
+		}
+		start = port
+		end = port
+	} else {
+		var err error
+		if start, err = strconv.Atoi(parts[0]); err != nil {
+			return nil, errors.Annotatef(err, "invalid port %q", parts[0])
+		}
+		if end, err = strconv.Atoi(parts[1]); err != nil {
+			return nil, errors.Annotatef(err, "invalid port %q", parts[1])
+		}
+	}
+
+	result := PortRange{
+		FromPort: start,
+		ToPort:   end,
+	}
+	return &result, nil
+}
+
+// ParsePortRanges splits the provided string on commas and extracts a
+// PortRange from each part of the split string. Whitespace is ignored.
+// Example strings: "80/tcp", "80,443,1234/udp", "123-456, 25/tcp".
+func ParsePortRanges(portRangesStr string) ([]PortRange, error) {
+	var portRanges []PortRange
+	for _, portRangeStr := range strings.Split(portRangesStr, ",") {
+		portRange, err := ParsePortRange(strings.TrimSpace(portRangeStr))
+		if err != nil {
+			return portRanges, errors.Trace(err)
+		}
+		portRanges = append(portRanges, *portRange)
+	}
+	return portRanges, nil
 }
